Swapped namespaces, moved dockerfile
