+++ AuthTests/UnitTests.cs
@@ -1,0 +1,7 @@
+namespace AuthTests
+{
+    public class UnitTests
+    {
+        
+    }
+}
