+++ AuthenticationAPI/AuthHelpers/HasScopeHandler.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
-namespace KitchenWeb.Helpers
+namespace AuthenticationAPI.AuthHelpers
 {
     public class HasScopeHandler : AuthorizationHandler<HasScopeRequirement>
     {
