+++ AuthenticationAPI/Controllers/AuthenticationController.cs
@@ -5,7 +5,8 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using Service.Authenticators;
+using AuthenticationAPI.AuthHelpers;
+
 
 namespace AuthenticationAPI.Controllers
 {
